requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.18.0
google-generativeai>=0.5.0
